+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/auth/session.ts
@@ -5,7 +5,7 @@ import { type AuthUserData } from '../server/auth/user.js';
 
 import { auth } from "./lucia.js";
 import type { Session } from "lucia";
-import { throwInvalidCredentialsError } from "./utils.js";
+import { createInvalidCredentialsError } from "./utils.js";
 
 import { prisma } from 'wasp/server';
 import { createAuthUserData } from "../server/auth/user.js";
@@ -65,7 +65,7 @@ async function getAuthUserData(userId: User['id']): Promise<AuthUserData> {
     })
 
   if (!user) {
-    throwInvalidCredentialsError()
+    throw createInvalidCredentialsError()
   }
 
   return createAuthUserData(user);