+++ waspc/e2e-test/test-outputs/waspJob-golden/waspJob/.wasp/out/sdk/wasp/server/jobs/simpleJob.ts
@@ -6,12 +6,14 @@ const entities = {
 }
 
 // PUBLIC API
-export type MySpecialJob<Input extends JSONObject, Output extends JSONValue | void> = JobFn<Input, Output, typeof entities>
+export type SimpleJob<Input extends JSONObject, Output extends JSONValue | void> = JobFn<Input, Output, typeof entities>
+
+const jobSchedule = null
 
 // PUBLIC API
-export const mySpecialJob = createJobDefinition({
-  jobName: 'mySpecialJob',
+export const simpleJob = createJobDefinition({
+  jobName: 'simpleJob',
   defaultJobOptions: {},
-  jobSchedule: null,
+  jobSchedule,
   entities,
 })