+++ waspc/cli/src/Wasp/Cli/Command/CreateNewProject.hs
@@ -15,14 +15,15 @@ import Wasp.Cli.Command.CreateNewProject.ProjectDescription
   ( NewProjectDescription (..),
     obtainNewProjectDescription,
   )
-import Wasp.Cli.Command.CreateNewProject.StarterTemplates
-  ( DirBasedTemplateMetadata (_path),
-    StarterTemplate (..),
-    getStarterTemplates,
-    getTemplateStartingInstructions,
-  )
+import Wasp.Cli.Command.CreateNewProject.StarterTemplates.FeaturedStarterTemplates (getFeaturedStarterTemplates)
 import Wasp.Cli.Command.CreateNewProject.StarterTemplates.GhRepo (createProjectOnDiskFromGhRepoTemplate)
 import Wasp.Cli.Command.CreateNewProject.StarterTemplates.Local (createProjectOnDiskFromLocalTemplate)
+import Wasp.Cli.Command.CreateNewProject.StarterTemplates.StarterTemplate
+  ( StarterTemplate (..),
+    WaspAppAiGenerator (WaspAI),
+    getTemplateName,
+    getTemplateStartingInstructions,
+  )
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import qualified Wasp.Message as Msg
 import qualified Wasp.Util.Terminal as Term
@@ -31,9 +32,8 @@ import qualified Wasp.Util.Terminal as Term
 createNewProject :: Arguments -> Command ()
 createNewProject args = do
   newProjectArgs <- parseNewProjectArgs args & either Common.throwProjectCreationError return
-  let starterTemplates = getStarterTemplates
-
-  newProjectDescription <- obtainNewProjectDescription newProjectArgs starterTemplates
+  let featuredStarterTemplates = getFeaturedStarterTemplates
+  newProjectDescription <- obtainNewProjectDescription newProjectArgs featuredStarterTemplates
 
   createProjectOnDisk newProjectDescription
   liftIO $ printGettingStartedInstructionsForProject newProjectDescription
@@ -46,20 +46,21 @@ createProjectOnDisk
       _template = template,
       _absWaspProjectDir = absWaspProjectDir
     } = do
-    cliSendMessageC $ Msg.Start $ "Creating your project from the \"" ++ show template ++ "\" template..."
+    cliSendMessageC $ Msg.Start $ "Creating your project from the \"" ++ getTemplateName template ++ "\" template..."
     case template of
-      GhRepoStarterTemplate ghRepoRef metadata ->
-        createProjectOnDiskFromGhRepoTemplate absWaspProjectDir projectName appName ghRepoRef $ _path metadata
-      LocalStarterTemplate metadata ->
-        liftIO $ createProjectOnDiskFromLocalTemplate absWaspProjectDir projectName appName $ _path metadata
-      AiGeneratedStarterTemplate ->
-        AI.createNewProjectInteractiveOnDisk absWaspProjectDir appName
+      GhRepoStarterTemplate ghRepoRef tmplDirPath _ ->
+        createProjectOnDiskFromGhRepoTemplate absWaspProjectDir projectName appName ghRepoRef tmplDirPath
+      LocalStarterTemplate tmplDirPath _ ->
+        liftIO $ createProjectOnDiskFromLocalTemplate absWaspProjectDir projectName appName tmplDirPath
+      AiGeneratedStarterTemplate waspAppAiGenerator _ ->
+        case waspAppAiGenerator of
+          WaspAI -> AI.createNewProjectInteractiveOnDisk absWaspProjectDir appName
 
 -- | This function assumes that the project dir was created inside the current working directory.
 printGettingStartedInstructionsForProject :: NewProjectDescription -> IO ()
 printGettingStartedInstructionsForProject projectDescription = do
   let projectDirName = init . SP.toFilePath . SP.basename $ _absWaspProjectDir projectDescription
-  let instructions = getTemplateStartingInstructions projectDirName $ _template projectDescription
+  let instructions = getTemplateStartingInstructions (_template projectDescription) projectDirName
   putStrLn $ Term.applyStyles [Term.Green] $ "Created new Wasp app in ./" ++ projectDirName ++ " directory!"
   putStrLn ""
   putStrLn instructions